+++ fcli-core/fcli-aviator-common/src/main/java/com/fortify/cli/aviator/fpr/remediation/classifier/PendingAppliedChange.java
@@ -10,24 +10,10 @@
  * herein. The information contained herein is subject to change
  * without notice.
  */
-package com.fortify.cli.aviator.fpr.remediation.model;
+package com.fortify.cli.aviator.fpr.remediation.classifier;
 
-import java.util.List;
+import java.nio.file.Path;
 
-public final class Remediation {
-    private final String instanceId;
-    private final List<FileChange> fileChanges;
-
-    public Remediation(String instanceId, List<FileChange> fileChanges) {
-        this.instanceId = instanceId;
-        this.fileChanges = fileChanges;
-    }
-
-    public String instanceId() {
-        return instanceId;
-    }
-
-    public List<FileChange> fileChanges() {
-        return fileChanges;
-    }
+public record PendingAppliedChange(Path filePath, String instanceId, int lineFrom, int lineTo, int deltaLines,
+                                    String comparisonCode) {
 }
